requirements.txt
Flask
openai
Frozen-Flask
python-dotenv
flask-cors